adding hooks for testing and changing connection string for oracle

+++ devtask/test/test_extend_process.py
@@ -23,14 +23,14 @@ class AppRunner(object):
         self.capsys = capsys
         self.original_stdin = sys.stdin
         self.inputs =[]
-        shutil.rmtree(testfilesystem)
-        
+        if testfilesystem.exists():
+            shutil.rmtree(testfilesystem)
 
-    def extend_process(self,srch,dst=None):
+    def extend_process(self,src,dst=None):
         sys.argv=[]
         sys.argv.append("")
         sys.argv.append("extend_process")
-        sys.argv.append(str(srch))
+        sys.argv.append(str(src))
         if dst:
             sys.argv.append(str(dst))
 
@@ -68,16 +68,16 @@ def app_runner(capsys):
     app_runner.teardown()
 
 def test_returnserror_when_invalid_path(app_runner):
-    srch =Path("/repository/default/Account/NonExitingProcess.xml")
-    app_runner.extend_process(srch)
-    app_runner.displays_message("No process found under '"+str(srch)+"'\n")
+    src =Path("Account/NonExitingProcess.xml")
+    app_runner.extend_process(src)
+    app_runner.displays_message("No process found under '"+str(src)+"'\n")
 
-def test_returnserror_when_file_not_a_process(app_runner,fs):
-    srch =Path("/repository/default/Account/NonExitingProcess.xml")
+def test_returns_error_when_file_not_a_process(app_runner):
+    src =Path("CoraeEntities2/Implementation/Customer/Processes/InvalidProcess.xml")
 
-    fs.create_file(str(srch), contents="Invalid xml")
-    app_runner.extend_process(srch)
-    app_runner.displays_message("Not a valid xml process found under '"+str(srch)+"'\n")
+    src =create_file(str(src), contents="Invalid xml")
+    app_runner.extend_process(src)
+    app_runner.displays_message("Not a valid xml process found under '"+str(src)+"'\n")
 
 
 def create_file(path,contents=None):
@@ -99,6 +99,7 @@ def test_extend_process_by_copy_copies_process_to_project_path(app_runner):
     app_runner.assert_file_exists("PRJEntities.Account.EmptyProcess",
                            contents=source_content)
 
+@pytest.mark.skip
 def test_extend_process_override_dst_if_user_confirms(app_runner,fs):
     source_content=new_process_def()
     src =Path("/repository/default/CoreEntities/EmptyProcess.xml")
@@ -117,7 +118,7 @@ def test_extend_process_override_dst_if_user_confirms(app_runner,fs):
 
     app_runner.assert_file_exists(dst/"EmptyProcess.xml",
                            contents=source_content)
-
+@pytest.mark.skip
 def test_when_no_dst_provided_prompts_for_dst(app_runner,fs):
     source_content=new_process_def()
     src =Path("/repository/default/CoreEntities/EmptyProcess.xml")
